Add instructions for a bookmarklet approach
I love this plugin and everything it stand for, but don't always have the option of installing it. On another level it is comforting to know that there's no extra work in the background to decide whether or not to activate. Even if I have to occasionally click to get paste back.  

✌️